refactoring and reorganisation of code
