+++ src/components/Details.jsx
@@ -45,7 +45,7 @@ const Details = ({ titleValue, descriptionValue, handleTitleChange, handleDescri
             </svg>
                 {showTooltip && (
                     <div className="absolute bg-black text-white text-xs rounded px-2 py-1 z-50">
-                        This is a lengthy test string tooltip for the description in order to test the textbox wrapping and size. I am extending this test string until the end of the page to see if it fits or if it wraps.
+                        Please input a short description of your project. This could include things like an overall purpose.
                     </div>
                 )}
                 </span>
