+++ src/components/MultiLineTextField/FuncRemoveButton.jsx
@@ -1,0 +1,20 @@
+import React from 'react';
+
+const RemoveButton = ({ onClick, className }) => {
+  return (
+    <button
+      type="button"
+      onClick={onClick}
+      className={`absolute right-1 ${className} cursor-pointer p-1 px-3 border-none focus:outline-none bg-white`}
+    >
+      <svg height="24px" width="24px" version="1.1" id="Capa_1" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 26 26">
+<g>
+	<path fill="#030104" d="M21.125,0H4.875C2.182,0,0,2.182,0,4.875v16.25C0,23.818,2.182,26,4.875,26h16.25   C23.818,26,26,23.818,26,21.125V4.875C26,2.182,23.818,0,21.125,0z M18.78,17.394l-1.388,1.387c-0.254,0.255-0.67,0.255-0.924,0   L13,15.313L9.533,18.78c-0.255,0.255-0.67,0.255-0.925-0.002L7.22,17.394c-0.253-0.256-0.253-0.669,0-0.926l3.468-3.467   L7.221,9.534c-0.254-0.256-0.254-0.672,0-0.925l1.388-1.388c0.255-0.257,0.671-0.257,0.925,0L13,10.689l3.468-3.468   c0.255-0.257,0.671-0.257,0.924,0l1.388,1.386c0.254,0.255,0.254,0.671,0.001,0.927l-3.468,3.467l3.468,3.467   C19.033,16.725,19.033,17.138,18.78,17.394z"/>
+</g>
+</svg>
+
+    </button>
+  );
+};
+
+export default RemoveButton;
